backend/requirements.txt
# --- Web framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9       # required for UploadFile / Form

# --- Settings ---
pydantic-settings>=2.2.0

# --- OpenAI ---
openai>=1.30.0
python-dotenv>=1.0.0

# --- Whisper (local model) ---
openai-whisper>=20231117
nest-asyncio>=1.6.0
